Api passport

+++ routes/api.php
@@ -17,7 +17,7 @@ Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::group(['middleware' => 'cors'], function(){
+Route::group(['middleware' => ['cors', 'auth:api' ] ], function(){
     //eta.test/api/allData
     Route::get('/allData','ApiGetController@allData');
 
@@ -32,30 +32,55 @@ Route::group(['middleware' => 'cors'], function(){
 
     ///////////////////////////////////////////////////////////////////
     //eta.test/api/newSite
-    Route::post('/newSite','ApiPostController@newSite');
+    Route::post('/newSite','ApiPostController@newSite')->middleware(['scope:admin,coordinator']);
 
     //eta.test/api/newEquipmentClass
-    Route::post('/newEquipmentClass','ApiPostController@newEquipmentClass');
+    Route::post('/newEquipmentClass','ApiPostController@newEquipmentClass')->middleware(['scope:admin,coordinator']);
 
     //eta.test/api/newEquip
-    Route::post('/newEquip','ApiPostController@newEquip');
+    Route::post('/newEquip','ApiPostController@newEquip')->middleware(['scope:admin,coordinator']);
 
     ///////////////////////////////////////////////////////////////////
     //eta.test/api/updateSite/8
-    Route::post('/updateSite/{site}','ApiPostController@updateSite');
+    Route::post('/updateSite/{site}','ApiPostController@updateSite')->middleware(['scope:admin,coordinator']);
 
     //eta.test/api/updateEquipmentClass/3
-    Route::post('/updateEquipmentClass/{equipClass}','ApiPostController@updateEquipmentClass');
+    Route::post('/updateEquipmentClass/{equipClass}','ApiPostController@updateEquipmentClass')->middleware(['scope:admin,coordinator']);
 
     //eta.test/api/updateEquip/10
     Route::post('/updateEquip/{equip}','ApiPostController@updateEquip');
 
     //eta.test/api/deleteSite/11
-    Route::delete('/deleteSite/{site}', 'ApiPostController@deleteSite');
+    Route::delete('/deleteSite/{site}', 'ApiPostController@deleteSite')->middleware(['scope:admin']);
 
     //eta.test/api/deleteEquipmentClass/7
-    Route::delete('/deleteEquipmentClass/{equipClass}', 'ApiPostController@deleteEquipmentClass');
+    Route::delete('/deleteEquipmentClass/{equipClass}', 'ApiPostController@deleteEquipmentClass')->middleware(['scope:admin']);
 
     //eta.test/api/deleteEquip/8
-    Route::delete('/deleteEquip/{equip}', 'ApiPostController@deleteEquip');
+    Route::delete('/deleteEquip/{equip}', 'ApiPostController@deleteEquip')->middleware(['scope:admin']);
+
+});
+
+//////////////////////////////////////
+// Authentication
+Route::group([ 'prefix' => 'auth', 'middleware' => 'cors' ], function () {
+    //eta.test/api/auth/login
+    Route::post('login', 'AuthController@login')->name('login');
+
+    //eta.test/api/auth/refreshToken
+    Route::post('refreshToken', 'AuthController@refreshToken');
+    
+    Route::group([
+        'middleware' => 'auth:api'
+    ], function() {
+        //eta.test/api/auth/logout
+        Route::get('logout', 'AuthController@logout');
+
+        //eta.test/api/auth/user
+        Route::get('user', 'AuthController@user');
+
+        //eta.test/api/auth/appToken
+        Route::post('appToken', 'AuthController@getAppToken');
+
+    });
 });
